+++ internal/controller/discovery/core/loader_interface.go
