+++ internal/controller/discovery/core/types.go
@@ -1,7 +1,5 @@
 package core
 
-import "sigs.k8s.io/controller-runtime/pkg/client"
-
 // DiscoveredTarget represents a target discovered from an external source
 // before it is materialized as a Kubernetes Target CR
 type DiscoveredTarget struct {
@@ -21,11 +19,4 @@ type DiscoveryEvent int
 type DiscoveryMessage struct {
 	Target DiscoveredTarget
 	Event  DiscoveryEvent
-}
-
-// TargetManager consumes discovered targets and applies them to Kubernetes.
-type TargetManager struct {
-	client       client.Client
-	targetsource string
-	in           <-chan []DiscoveryMessage
 }
