+++ internal/controller/discovery/loaders.go
