--- internal/controller/targetsource/core/loaders_test.go
@@ -1,1 +1,0 @@
-package core
