--- internal/controller/targetsource/mapper_test.go
@@ -1,1 +1,0 @@
-package targetsource
